CDK-249: Update for recommendations on PR#8.
This alters the partitionAdded signature to use a String rather than a
StorageKey. This removes the need to alter StorageKey and to expose the
dirnameForValue method through the Accessor. In addition, the partition
logic is kept in FileSystemDataset and PartitionedDatasetWriter rather
than in FileSystemWriters. When calling partitionAdded, the
implementations now pass the String path component that was actually
used to create the writer.