CDK-342: Add descriptor builder methods for JSON partition strategies.
This adds methods that mirror the schema setting methods in the
DatasetDescriptor.Builder. In addition, if the partition strategy is not
explicitly set using those methods, then it is parsed from the schema.

This commit also removes the Schema from the parser. Partition
strategies can be defined independently, and consistency is checked in
the DatasetDescriptor. The Schema was only used to set the
IdentityFieldPartitioner source (and result) class, which is now Object.
The only place this was used specifically was in valueFromString, which
now requires the expected type.