Fix Hive MetaStore URI handling.
The hive.metastore.uris property is set in hive-site.xml, but the Loader
and HCatalog class were checking the property before creating a HiveConf
that reads the hive-site.xml file. This prevented the environment's
MetaStore URI from being loaded.

HiveConf is now required in the Loader to initialize URIs correctly, but
this caused test failures kite-data-mapreduce because the MR tasks did
not have Hive jars in the classpath. This commit adds dynamic resolution
of HiveConf so that the hive URIs will not be registered if it can't be
loaded without failing registration for all implementations.

This also updates the logic to fail rather than warn if a local
MetaStore is being used because the local MS is only for testing. To
avoid this failure, this commit introduces the property,
kite.allow.test-metastore to override the new behavior and adds a
hadoop-site.xml file to the test resources that adds it to the test
environment. (Use hadoop-site.xml because hive-site.xml breaks Hive.)