CDK-192: Fix CopyTask to work with local datasets.
Crunch's MemPipeline won't write to MapReduceTargets, so the CopyTask
has to write to the output View manually. Crunch is still used to
materialize the incoming collection so that DoFn transforms can be used
later.

This also updates the csv-import command to use the CopyTask. However,
because the CSV dataset is not permanent, it can't be used with
MapReduce and the copy is forced to run locally.