CDK-664: Fix URI encoding consistency in URIBuilder.
This updates the URIBuilder for consistent construction and parsing of
encoded values in the path and query components of a URI. Previously,
path and query strings were accessed with the getPath and getQuery
methods, which will decode encoded values. This breaks URIs where
delimiters are encoded in the data (a%2Cb => a,b => a or b). The fix is
to use the "raw" accessors that do not decode values.

It is also necessary to construct URIs without encoding the path and
query values. To do this, only the single-String URI constructor or
URI.create can be used. The URI constructors that accept URI parts and
construct a URI will encode delimiter values and double-encode already
encoded values (a%2Cb,c => a%252Cb%2Cc).